+++ tests/testthat/tests.R
@@ -28,3 +28,33 @@ test_that("addPlots", {
     Plots,debug=F,height=4)},
   paste(tempdir(),'/resultPlots.docx',sep = ''))
 })
+
+test_that("addPlots2", {
+  library(ggplot2)
+  outpath<-paste(tempdir(),'/resultPlots2.docx',sep = '')
+  Plots<-list(plot1=function(){plot(hp~wt,data=mtcars,col=cyl)},
+              plot2=(ggplot(mtcars,aes(x=wt,y=hp,col=as.factor(cyl)))+geom_point()))
+  resFile<-addPlots(
+    paste(examplePath(),'templates/templatePlots.docx',sep = ''),
+    outpath,
+    Plots,debug=F,height=4, width=c(3,6))
+
+  expect_equal(resFile, outpath)
+})
+
+
+
+test_that("addPlots3", {
+  library(ggplot2)
+  outpath<-paste(tempdir(),'/resultPlots2.docx',sep = '')
+  Plots<-list(plot1=function(){plot(hp~wt,data=mtcars,col=cyl)},
+              plot2=(ggplot(mtcars,aes(x=wt,y=hp,col=as.factor(cyl)))+geom_point()))
+  resFile<-addPlots(docxIn= officer::read_docx(path =paste(examplePath(),'templates/templatePlots.docx',sep = '')),
+                    docxOut=NA,
+                    Plots,debug=F,height=4, width=c(3,6))
+
+  print(resFile, target = outpath)
+
+  expect_equal(outpath, outpath)
+})
+